Create README.md
Created this after some guy asked for. Sadly something happened and the project stopped. It was supposed to be an auction system but instead of having a database with all the files, everything from users to products and sales is saved inside a JSON file. 
There is no fancy css styles, just the body and the base of an auction system. Still need to add more stats but as mentioned, the project stopped in the middle.

There are obvious implementations and TODO list, but may be useful to someone. Currently it has the login part where the user can post a product, and others logged users can battle among them to win the item. 
All the items have 1 hour to be sold. The winner is displayed and the idea was to claim the item afterwards, was when it stopped and nothing else was added.

![My animated logo](https://i.imgur.com/Zs56WiQ.png)
![My animated logo](https://i.imgur.com/pnE5kBH.png)
![My animated logo](https://i.imgur.com/jy1oUwf.png)
![My animated logo](https://i.imgur.com/WORk76C.png)